- merged last_inserted_ids() fix from trunk [changeset:2743]

+++ test/sql/query.py
@@ -36,25 +36,39 @@ class QueryTest(PersistTest):
         users.insert().execute(user_id = 7, user_name = 'jack')
         assert users.count().scalar() == 1
     
-    @testbase.unsupported('sqlite')
+    def testupdate(self):
+
+        users.insert().execute(user_id = 7, user_name = 'jack')
+        assert users.count().scalar() == 1
+
+        users.update(users.c.user_id == 7).execute(user_name = 'fred')
+        assert users.select(users.c.user_id==7).execute().fetchone()['user_name'] == 'fred'
+
     def test_lastrow_accessor(self):
         """test the last_inserted_ids() and lastrow_has_id() functions"""
-        
+
         def insert_values(table, values):
+            """insert a row into a table, return the full list of values INSERTed including defaults
+            that fired off on the DB side.  
+            
+            detects rows that had defaults and post-fetches.
+            """
+            
             result = table.insert().execute(**values)
             ret = values.copy()
-            
+
             for col, id in zip(table.primary_key, result.last_inserted_ids()):
                 ret[col.key] = id
-                
+
             if result.lastrow_has_defaults():
                 criterion = and_(*[col==id for col, id in zip(table.primary_key, result.last_inserted_ids())])
                 row = table.select(criterion).execute().fetchone()
                 ret.update(row)
             return ret
-            
-        for table, values, assertvalues in [
+
+        for supported, table, values, assertvalues in [
             (
+                {'unsupported':['sqlite']},
                 Table("t1", metadata, 
                     Column('id', Integer, primary_key=True),
                     Column('foo', String(30), primary_key=True)),
@@ -62,6 +76,7 @@ class QueryTest(PersistTest):
                 {'id':1, 'foo':'hi'}
             ),
             (
+                {'unsupported':['sqlite']},
                 Table("t2", metadata, 
                     Column('id', Integer, primary_key=True),
                     Column('foo', String(30), primary_key=True),
@@ -70,21 +85,43 @@ class QueryTest(PersistTest):
                 {'foo':'hi'},
                 {'id':1, 'foo':'hi', 'bar':'hi'}
             ),
-            
+            (
+                {'unsupported':[]},
+                Table("t3", metadata, 
+                    Column("id", String(40), primary_key=True),
+                    Column('foo', String(30), primary_key=True),
+                    Column("bar", String(30))
+                    ),
+                    {'id':'hi', 'foo':'thisisfoo', 'bar':"thisisbar"},
+                    {'id':'hi', 'foo':'thisisfoo', 'bar':"thisisbar"}
+            ),
+            (
+                {'unsupported':[]},
+                Table("t4", metadata, 
+                    Column('id', Integer, primary_key=True),
+                    Column('foo', String(30), primary_key=True),
+                    Column('bar', String(30), PassiveDefault('hi'))
+                ),
+                {'foo':'hi', 'id':1},
+                {'id':1, 'foo':'hi', 'bar':'hi'}
+            ),
+            (
+                {'unsupported':[]},
+                Table("t5", metadata, 
+                    Column('id', String(10), primary_key=True),
+                    Column('bar', String(30), PassiveDefault('hi'))
+                ),
+                {'id':'id1'},
+                {'id':'id1', 'bar':'hi'},
+            ),
         ]:
+            if testbase.db.name in supported['unsupported']:
+                continue
             try:
                 table.create()
-                assert insert_values(table, values) == assertvalues
+                assert insert_values(table, values) == assertvalues, repr(values) + " " + repr(assertvalues)
             finally:
                 table.drop()
-            
-    def testupdate(self):
-
-        users.insert().execute(user_id = 7, user_name = 'jack')
-        assert users.count().scalar() == 1
-
-        users.update(users.c.user_id == 7).execute(user_name = 'fred')
-        assert users.select(users.c.user_id==7).execute().fetchone()['user_name'] == 'fred'
 
     def testrowiteration(self):
         users.insert().execute(user_id = 7, user_name = 'jack')
